refactor(ui): extract tabRenderer subrenderers
tabRenderer.js 從 561 行降為 353 行（-37%）：

- 新增 modules/ui/tab/tabListeners.js (170 行)：
  把 initTabListeners 整段與 5 個 event delegation 搬出。
  內部自治持有 listenersInitialized + listenerAbortController，
  透過 deps { getTabsCache, getAddToGroupCallback } 解耦，
  以避免從 tabRenderer 帶閉包進來。

- 新增 modules/ui/tab/splitViewRenderer.js (77 行)：
  export buildSplitTabsMap(tabs) + createSplitOrTabRenderer(ctx)
  factory。renderTabsAndGroups 內的 split-view 邏輯與 splitTabsMap
  建立都委派出去。

createTabElement / updateTabElement / updateGroupHeaderElement 保留
主檔（單一函式 50-80 行、耦合度低，硬拆會增加 module 切換成本而
無收益）。完整四檔拆分留待未來補完 unit test 後再評估。

Public API 完全不變：renderTabsAndGroups / resetTabListeners /
getTabCache / getTabElementsCache / getGroupHeaderElementsCache /
createTabElement 簽名與行為一致，外部呼叫端 (searchManager.js /
uiManager.js) 無需改動。

驗證：21 puppeteer tests 通過（happy_path_sidepanel_load /
tab_close / tab_switch / tab_group_toggle / add_to_group /
keyboard_a11y / scroll_position_on_tab_click）。
