+++ modules/ui/tabRenderer.js
@@ -1,50 +1,35 @@
 import * as api from '../apiManager.js';
 import { ADD_TO_GROUP_ICON_SVG, BOOKMARK_ICON_SVG } from '../icons.js';
 import { tabListContainer } from './elements.js';
-import { showContextMenu } from './contextMenuManager.js';
 import { GROUP_COLORS, hexToRgba } from './groupColors.js';
 import { reconcileDOM } from '../utils/domUtils.js';
+import { initTabListeners, resetTabListeners as resetListeners } from './tab/tabListeners.js';
+import { buildSplitTabsMap, createSplitOrTabRenderer } from './tab/splitViewRenderer.js';
 
 /**
- * Module-level state for event delegation.
- * These are intentionally module-scoped to persist across renders.
+ * Module-level state for event delegation and render-cycle caching.
+ * Caches are reassigned per render; listeners read them via getter to see latest.
  */
 
-/** @type {Function|null} Callback for adding a tab to a group, set during renderTabsAndGroups */
+/** @type {Function|null} Set per renderTabsAndGroups call, read by listeners on click. */
 let currentAddToGroupCallback = null;
 
-/** @type {boolean} Flag to ensure event listeners are only initialized once */
-let listenersInitialized = false;
-
-/** @type {Map<number, chrome.tabs.Tab>} Cache of tab objects for event delegation handlers */
+/** @type {Map<number, chrome.tabs.Tab>} */
 let tabsCache = new Map();
-
-/** @type {Map<number, HTMLElement>} Cache of tab DOM elements for performance optimization */
+/** @type {Map<number, HTMLElement>} */
 let tabElementsCache = new Map();
-
-/** @type {Map<number, HTMLElement>} Cache of group header DOM elements for performance optimization */
+/** @type {Map<number, HTMLElement>} */
 let groupHeaderElementsCache = new Map();
-
-/** @type {Map<number, HTMLElement>} Cache of group content container DOM elements for performance optimization */
+/** @type {Map<number, HTMLElement>} */
 let groupContentElementsCache = new Map();
-
-/** @type {Map<number, HTMLElement>} Cache of split group container DOM elements for performance optimization */
+/** @type {Map<number, HTMLElement>} */
 let splitGroupElementsCache = new Map();
 
-/** @type {AbortController|null} Controller to abort event listeners when resetting */
-let listenerAbortController = null;
-
 /**
- * Resets the tab list listeners and module state.
- * Useful for testing, hot-reload, or when the container element is replaced.
- * @returns {void}
+ * Resets listeners and all render caches. For tests / container replacement.
  */
 export function resetTabListeners() {
-    if (listenerAbortController) {
-        listenerAbortController.abort();
-        listenerAbortController = null;
-    }
-    listenersInitialized = false;
+    resetListeners();
     currentAddToGroupCallback = null;
     tabsCache = new Map();
     tabElementsCache = new Map();
@@ -75,161 +60,6 @@ export function getTabElementsCache() {
  */
 export function getGroupHeaderElementsCache() {
     return groupHeaderElementsCache;
-}
-
-function initTabListeners(container) {
-    if (listenersInitialized) return;
-    listenersInitialized = true;
-
-    // Track input method: only auto-scroll on focusin for keyboard navigation.
-    // Mouse clicks skip scrollIntoView to prevent scroll jump after DOM reconciliation.
-    let lastInputWasKeyboard = true;
-
-    // Create AbortController for cleanup support
-    listenerAbortController = new AbortController();
-    const { signal } = listenerAbortController;
-
-    // Helper to find tab data from cache
-    const getTabFromElement = (element) => {
-        const tabItem = element.closest('.tab-item');
-        if (!tabItem) return null;
-        const tabId = parseInt(tabItem.dataset.tabId);
-        return tabsCache.get(tabId);
-    };
-
-    // Helper to find group data
-    const getGroupData = (element) => {
-        const header = element.closest('.tab-group-header');
-        if (!header) return null;
-        return {
-            id: parseInt(header.dataset.groupId),
-            collapsed: header.dataset.collapsed === 'true',
-            element: header
-        };
-    };
-
-    // --- Click Delegation ---
-    container.addEventListener('click', async (e) => {
-        // Handle Button Clicks
-        const button = e.target.closest('button');
-        if (button) {
-            e.stopPropagation();
-            const action = button.dataset.action;
-            const tab = getTabFromElement(button);
-
-            if (!tab) return;
-
-            if (action === 'close') {
-                api.removeTab(tab.id);
-            } else if (action === 'add-to-group') {
-                if (currentAddToGroupCallback) {
-                    currentAddToGroupCallback(tab.id);
-                }
-            } else if (action === 'add-to-bookmark') {
-                const modal = await import('../modalManager.js');
-                const result = await modal.showAddToBookmarkDialog({
-                    name: tab.title,
-                    url: tab.url
-                });
-                if (result) {
-                    await api.createBookmark(result);
-                }
-            }
-            return;
-        }
-
-        // Handle Group Header Click
-        const groupData = getGroupData(e.target);
-        if (groupData) {
-            const header = groupData.element;
-            const content = header.nextElementSibling;
-            const arrow = header.querySelector('.tab-group-arrow');
-
-            const newCollapsedState = !groupData.collapsed;
-            content.style.display = newCollapsedState ? 'none' : 'block';
-            arrow.textContent = newCollapsedState ? '▶' : '▼';
-            header.setAttribute('aria-expanded', (!newCollapsedState).toString());
-            header.dataset.collapsed = newCollapsedState ? 'true' : 'false';
-
-            api.updateTabGroup(groupData.id, { collapsed: newCollapsedState });
-            return;
-        }
-
-        // Handle Tab Click (Activation)
-        const tab = getTabFromElement(e.target);
-        if (tab) {
-            api.updateTab(tab.id, { active: true });
-            api.updateWindow(tab.windowId, { focused: true });
-        }
-    }, { signal });
-
-    // --- Mouse input tracking ---
-    container.addEventListener('mousedown', () => {
-        lastInputWasKeyboard = false;
-    }, { signal });
-
-    // --- Keydown Delegation ---
-    container.addEventListener('keydown', (e) => {
-        lastInputWasKeyboard = true;
-        const tabItem = e.target.closest('.tab-item');
-        const groupHeader = e.target.closest('.tab-group-header');
-
-        if (tabItem) {
-            // Prevent handling if focus is on a button inside the tab
-            if (e.target.tagName === 'BUTTON') return;
-
-            const tab = getTabFromElement(tabItem);
-            if (!tab) return;
-
-            if (e.key === 'Enter') {
-                e.preventDefault();
-                api.updateTab(tab.id, { active: true });
-                api.updateWindow(tab.windowId, { focused: true });
-            } else if (e.key === 'Delete') {
-                e.preventDefault();
-                e.stopPropagation();
-                api.removeTab(tab.id);
-            } else if (e.key === ' ') {
-                e.preventDefault();
-                const existingMenu = document.querySelector('.custom-context-menu');
-                if (existingMenu) {
-                    existingMenu.remove();
-                } else {
-                    const rect = tabItem.getBoundingClientRect();
-                    showContextMenu(rect.left + 20, rect.bottom, tab, tabItem);
-                }
-            }
-        } else if (groupHeader) {
-            if (e.key === 'Enter' || e.key === ' ') {
-                e.preventDefault();
-                groupHeader.click();
-            }
-        }
-    }, { signal });
-
-    // --- Context Menu Delegation ---
-    container.addEventListener('contextmenu', (e) => {
-        const tab = getTabFromElement(e.target);
-        const tabItem = e.target.closest('.tab-item');
-        if (tab && tabItem) {
-            e.preventDefault();
-            showContextMenu(e.clientX, e.clientY, tab, tabItem);
-        }
-    }, { signal });
-
-    // --- Focus In (scrolling) ---
-    // Only auto-scroll when navigating via keyboard, not on mouse click.
-    // Without this guard, DOM reconciliation after onActivated triggers focusin,
-    // which calls scrollIntoView and unexpectedly jumps the scroll position.
-    container.addEventListener('focusin', (e) => {
-        if (!lastInputWasKeyboard) return;
-        const tabItem = e.target.closest('.tab-item');
-        if (tabItem) {
-            setTimeout(() => {
-                tabItem.scrollIntoView({ block: 'nearest', behavior: 'smooth' });
-            }, 0);
-        }
-    }, { signal });
 }
 
 function updateTabElement(tabItem, tab) {
@@ -396,9 +226,10 @@ export function createTabElement(tab, { onAddToGroupClick }) {
 }
 
 export function renderTabsAndGroups(tabs, groups, { onAddToGroupClick }) {
-    if (!listenersInitialized) {
-        initTabListeners(tabListContainer);
-    }
+    initTabListeners(tabListContainer, {
+        getTabsCache: () => tabsCache,
+        getAddToGroupCallback: () => currentAddToGroupCallback,
+    });
     currentAddToGroupCallback = onAddToGroupClick;
 
     // Update cache with full tab objects
@@ -426,15 +257,7 @@ export function renderTabsAndGroups(tabs, groups, { onAddToGroupClick }) {
     const renderedTabIds = new Set();
     const topLevelChildren = [];
 
-    const splitTabsMap = new Map();
-    for (const tab of tabs) {
-        if (tab.splitViewId && tab.splitViewId > 0) {
-            if (!splitTabsMap.has(tab.splitViewId)) {
-                splitTabsMap.set(tab.splitViewId, []);
-            }
-            splitTabsMap.get(tab.splitViewId).push(tab);
-        }
-    }
+    const splitTabsMap = buildSplitTabsMap(tabs);
 
     // Helper to get or create tab element
     const getOrCreateTabElement = (tab) => {
@@ -448,44 +271,13 @@ export function renderTabsAndGroups(tabs, groups, { onAddToGroupClick }) {
         return tabElement;
     };
 
-    // Helper to render split groups
-    const renderSplitOrTab = (tab, targetArray) => {
-        if (renderedTabIds.has(tab.id)) return;
-
-        if (tab.splitViewId && tab.splitViewId > 0) {
-            // Find all tabs in the same split view context
-            const splitTabs = splitTabsMap.get(tab.splitViewId);
-
-            if (splitTabs && splitTabs.length > 1) {
-                // Get or create split group container (reuse from cache for DOM reconciliation)
-                let splitGroup = splitGroupElementsCache.get(tab.splitViewId);
-                if (!splitGroup) {
-                    splitGroup = document.createElement('div');
-                    splitGroup.className = 'tab-split-group';
-                }
-                newSplitGroupElementsCache.set(tab.splitViewId, splitGroup);
-
-                // Build children array and reconcile
-                const splitChildren = [];
-                splitTabs.forEach(splitTab => {
-                    const tabElement = getOrCreateTabElement(splitTab);
-                    tabElement.classList.add('in-split-view');
-                    splitChildren.push(tabElement);
-                    renderedTabIds.add(splitTab.id);
-                });
-                reconcileDOM(splitGroup, splitChildren);
-
-                targetArray.push(splitGroup);
-                return;
-            }
-        }
-
-        // Fallback to normal rendering if not split view or only 1 tab in split
-        const tabElement = getOrCreateTabElement(tab);
-        tabElement.classList.remove('in-split-view'); // Ensure it's clean if it was in split before
-        targetArray.push(tabElement);
-        renderedTabIds.add(tab.id);
-    };
+    const renderSplitOrTab = createSplitOrTabRenderer({
+        splitTabsMap,
+        splitGroupElementsCache,
+        newSplitGroupElementsCache,
+        renderedTabIds,
+        getOrCreateTabElement,
+    });
 
     for (const tab of tabs) {
         if (renderedTabIds.has(tab.id)) {
